move legacy types to separate file
